requirements.txt
djangorestframework==3.15.1
django==5.2.4
dj-rest-auth==7.0.1
django-allauth==0.61.1
djangorestframework-simplejwt==5.3.1
gunicorn
whitenoise
psycopg2-binary
dj-database-url
django-taggit>=5.0
django-cors-headers
python-decouple
Pillow
environ
django-environ
psycopg2-binary
dj-database-url
